requirements.txt
langchain
langchain-community
llama-parse
langchain-nomic
langchain-ollama
scikit-learn
streamlit
ollama
langchain-core
langchain_text_splitters
transformers
PyPDF2
faiss-cpu
torch
langgraph
Ipython
ipykernel
langchain-huggingface
sentence_transformers